*** empty log message ***

+++ src/ulib_std.cpp
@@ -218,7 +218,7 @@ float mtof(float f)
 {
     if (f <= -1500) return(0);
     else if (f > 1499) return(mtof(1499));
-    else return (8.17579891564 * exp(.0577622650 * f));
+    else return (float)(8.17579891564 * exp(.0577622650 * (double)f));
 }
 
 CK_DLL_FUNC( mtof_impl )
